+++ src/components/LoginPage/loginpage.scss
@@ -1,0 +1,41 @@
+
+
+    .loginpage{
+        background-color: green;
+        display: flex;
+        flex-direction: row;
+        justify-content: center;
+        align-items: center;
+        width: 100%;
+        height: 100vh;
+
+        .loginpage-form-container{
+            background-color: hotpink;
+            height: 100%;
+            width: 100%;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+        }
+        .loginpage-photo-container{
+            background-color: indigo;
+            width: 100%;
+            height: 100%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+
+            .background-image{
+                background-image: url('../../img/Login/abstract-login-image.png');
+                height: 90%;
+                width: 90%;
+                position: relative;
+                .text-image{
+                    position: absolute;
+                    left: -20%;
+                }
+            }
+
+        }
+    }
